% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvmcorr.R
\name{jcorr_model_simple}
\alias{jcorr_model_simple}
\title{Paralleled full Bayesian estimation for multivariate dyads data analysis using a
joint latent variable framework with linear correlation model.}
\usage{
jcorr_model_simple(formula_mean, formula_corr, data, mcmc_len, verbose = F)
}
\arguments{
\item{formula_mean}{an object of class "formula" similar to linear regression.
It is a symbolic description of the covariates set used in the mean model.}

\item{formula_corr}{an object of class "formula" similar to linear regression.
It is a symbolic description of the covariates set used in the correlation model.}

\item{data}{a list for data input. It contains binary response matrices,
covariate variables (See data_dictionary.txt for more details).}

\item{mcmc_len}{an integer specifies the length of MCMC draws.}

\item{verbose}{boolen variable, whether enable verbose mode.}
}
\value{
The function returns a list with the following components:
\describe{
\item{g_draws}{A matrix containing MCMC draws for coefficient parameters \eqn{\phi_\xi}.}
\item{b_tp_draws}{A matrix containing MCMC draws for coefficient parameters \eqn{\beta_{GP}}.}
\item{b_fp_draws}{A matrix containing MCMC draws for coefficient parameters \eqn{\beta_{RP}}.}
\item{u_tp_draws}{A matrix containing MCMC draws for coefficient parameters \eqn{\beta_{GF}}.}
\item{u_fp_draws}{A matrix containing MCMC draws for coefficient parameters \eqn{\beta_{RF}}.}
\item{sig_draws}{A matrix containing MCMC draws for variances of \eqn{\eta_1} and \eqn{\eta_2}.}
\item{beta_rho_draws}{A matrix containing MCMC draws for coefficients of correlations.}
\item{e1234_last}{A matrix containing MCMC draws for continuous random variables \eqn{\eta_1}-\eqn{\eta_4}.}
\item{xi_last}{A matrix containing MCMC draws for discrete random variables \eqn{\xi}.}
}
}
\description{
The program perform estimation for a general latent variable modelling framework.
This framework is intended for for the analysis of
multivariate binary data on dyads, which can also handle zero inflation and
non-equivalence of measurement.
}
\references{
Zhang, S., Kuha, J., & Steele, F. (2023). Modelling correlation matrices in multivariate data,
with application to reciprocity and complementarity of child-parent exchanges of support.
(https://arxiv.org/abs/2210.14751)
}
