% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvmcorr.R
\name{lvmcorr}
\alias{lvmcorr}
\title{An R package for joint structural equation modeling}
\usage{
lvmcorr(
  ytp,
  yfp,
  x_covariates,
  b_tp_init,
  b_fp_init,
  g_init,
  g_free_ind,
  sig2_tp_init,
  sig2_fp_init,
  rho_init,
  e_tp,
  e_fp,
  e12,
  xi,
  beta_tp,
  alpha_tp,
  beta_fp,
  alpha_fp,
  mcmc_len
)
}
\description{
An R package for joint structural equation modeling
}
